Create separate start scripts for the api server and persister
Both need to write out the encryption key and wait for postgres,
but after that the startup process is different.

We will specify the common "entrypoint" in the Dockerfile, but will
add the specific script to run in the deployment spec in the
openshift template.


(transferred from ManageIQ/topological_inventory-core@af509ce22a5d09ec80e84f3562ac5cf6335846bb)